Refactoring: GridCreatingConditionTriangle::FinishPolyLineDefiningCommand implementation is separated from gridcreatingconditiontriangle.cpp.
